WFLY-15784 Merge ejb3 ScheduleTimer and AutoTimer classes
