+++ ejb3/src/main/java/org/jboss/as/ejb3/timerservice/TimerServiceImpl.java
@@ -31,7 +31,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.Map;
@@ -64,7 +63,6 @@ import org.jboss.as.ejb3.subsystem.deployment.TimerServiceResource;
 import org.jboss.as.ejb3.timerservice.persistence.TimerPersistence;
 import org.jboss.as.ejb3.timerservice.persistence.database.DatabaseTimerPersistence;
 import org.jboss.as.ejb3.timerservice.schedule.CalendarBasedTimeout;
-import org.jboss.as.ejb3.timerservice.spi.ScheduleTimer;
 import org.jboss.as.ejb3.timerservice.spi.TimedObjectInvoker;
 import org.jboss.invocation.InterceptorContext;
 import org.jboss.msc.service.Service;
@@ -196,10 +194,15 @@ public class TimerServiceImpl implements TimerService, Service<TimerService> {
 
 
     public synchronized void activate() {
-        final List<ScheduleTimer> timers = new ArrayList<ScheduleTimer>();
-        for (Map.Entry<Method, List<AutoTimer>> entry : autoTimers.entrySet()) {
-            for (AutoTimer timer : entry.getValue()) {
-                timers.add(new ScheduleTimer(entry.getKey(), timer.getScheduleExpression(), timer.getTimerConfig()));
+        final List<AutoTimer> timers;
+        if (autoTimers.isEmpty()) {
+            timers = Collections.emptyList();
+        } else {
+            timers = new ArrayList<>();
+            for (Map.Entry<Method, List<AutoTimer>> entry : autoTimers.entrySet()) {
+                for (AutoTimer timer : entry.getValue()) {
+                    timers.add(new AutoTimer(timer.getScheduleExpression(), timer.getTimerConfig(), entry.getKey()));
+                }
             }
         }
         // restore the timers
@@ -723,14 +726,11 @@ public class TimerServiceImpl implements TimerService, Service<TimerService> {
      * All such restored timers will be schedule for their next timeouts.
      * </p>
      *
-     * @param autoTimers
+     * @param newAutoTimers
      */
-    public void restoreTimers(final List<ScheduleTimer> autoTimers) {
+    public void restoreTimers(final List<AutoTimer> newAutoTimers) {
         // get the persisted timers which are considered active
         List<TimerImpl> restorableTimers = this.getActivePersistentTimers();
-
-        //timers are removed from the list as they are loaded
-        final List<ScheduleTimer> newAutoTimers = new LinkedList<ScheduleTimer>(autoTimers);
 
         if (EJB3_TIMER_LOGGER.isDebugEnabled()) {
             EJB3_TIMER_LOGGER.debug("Found " + restorableTimers.size() + " active persistentTimers for timedObjectId: "
@@ -744,9 +744,9 @@ public class TimerServiceImpl implements TimerService, Service<TimerService> {
                 CalendarTimer calendarTimer = (CalendarTimer) activeTimer;
                 boolean found = false;
                 //so we know we have an auto timer. We need to try and match it up with the auto timers.
-                ListIterator<ScheduleTimer> it = newAutoTimers.listIterator();
+                ListIterator<AutoTimer> it = newAutoTimers.listIterator();
                 while (it.hasNext()) {
-                    ScheduleTimer timer = it.next();
+                    AutoTimer timer = it.next();
                     if (doesTimeoutMethodMatch(calendarTimer.getTimeoutMethod(), timer.getMethod())) {
 
                         //the timers have the same method.
@@ -785,7 +785,7 @@ public class TimerServiceImpl implements TimerService, Service<TimerService> {
             EJB3_TIMER_LOGGER.debugv("Started timer: {0}", activeTimer);
         }
 
-        for (ScheduleTimer timer : newAutoTimers) {
+        for (AutoTimer timer : newAutoTimers) {
             this.loadAutoTimer(timer.getScheduleExpression(), timer.getTimerConfig(), timer.getMethod());
         }
